Move cursor to end if user clicks outside any node
If the user clicks on the field, but the click isn't caught by any
individual node, move the cursor to the rightmost position.

Although measuring mouse position would be strictly better, mostly when
clicking on the field but not on a specific node I expect the cursor to
move to the end of the input.

Also now supports clicking on Expressions outside of their child nodes.
So, for example, clicking in a numerator will move to the end of the
numerator, rather than previous behaviour of that click being caught by
the DivisionNode and the cursor being moved right of the DivisionNode
instead.