Add tests, restructure files as necessary
Jest was installed to facilitate testing, which in turn necessitated the
installation of Yarn. Babel was installed to allow for ES6 testing.

Although the project has been initialised through Yarn, it's as yet
unpublished, and should still be installed simply by copying the source
files.

Due to testing limitations, math-input.js was split into math-input.js,
which handles only the WebComponent, and math-node.js, which handles all
of the MathNodes themselves.