Fix bug incorrectly parsing characters as pi
I forgot a set of parentheses in the pi-matching regex, which meant that
if there was a pi character in an expression, any symbol before it but
with lower priority would be parsed as pi. (Priority in the _parse()
function is rather arbitrary, as only symbols where there is possible
confusion, like checking for 'sin' before 's', is carefully ordered.)

This went unnoticed before since usually pi is in brackets (which masks
it from consideration), or preceded by numbers (which are higher
priority).