Allow <math-input> field to work in MS Edge
The <math-input> field previously didn't work in Edge because the
browser lacks support for WebComponents.

Added a polyfill for support, and altered the code slightly: previously
the actual <input> form element which stores the MathML value of the
expression typed was a child of the <math-input> element (on the real
DOM, not the shadow). For some reason this wasn't working in Edge, so
placed it as a sibling to the <math-input>.

This shouldn't alter any behaviour as the <input>'s ID and other
characteristics remain the same, and it will still be a descendant of
all the same ancestor elements.

Updated readme to discuss dependency installation.