Update README, refer to mathinput by ID not name
Previously in the README, the mathinput element was referred to in
Javascript using getElementsByName():
    document.getElementsByName('fieldName')[0]...

The [0] wasn't explained, but it was necessary because the field, on
instantiation, creates a hidden input with the same name to contain the
MathML, so that the field could be used in a form without additional
Javascript calls.

This broke in the previous commit because I started placing the hidden
input before the <math-input> element, rather than after, in the DOM.
But it was just inherently unstable anyway because that sort of detail
shouldn't be relevant.

Changed all references in the README to use ID instead, and added a note
warning about using getElementsByName().