Change method of setting math-input using MathML
Previously, the MathML would be converted to a list of MathNodes by a
static function, and would then be appended to the input field.

This failed when I tried to handle division, however, because the
functions for adding a numerator or denominator to a division node
assume that the DivisionNode is attached to its parent and, ultimately,
the DOM.

This probably isn't ideal, but it's too big a thing to change, so
instead I've made the functions nonstatic, and instead the MathML string
is passed to the ExpressionNode to which it's going to be appended.

I haven't added any new tests, but all the previous tests pass.