Don't show times symbol in preset values
There's no need to show the times symbol itself in a multiplication,
it's implied. There may be edge cases I've missed where it's ambiguous,
but none that I can think of. Now, when a value is preset using the
value attribute, any times symbols won't be shown.

All tests still pass, indicating this is reliably producing the correct
output, except one which checked the visual presentation of a
multiplication, which obviously failed when the presentation changed.
It's been adapted.