Allow backspacing of the field using a button
Added the `action` attribute to the <math-input> field, which can be
used to perform non-insert actions. Currently the only implemented
action is `backspace`, which does exactly the same thing as pressing the
backspace key.

Added to better support mobile use, because getting the system keyboard
to appear when using the field on phones is proving difficult.